backend/requirements.txt
Flask>=3.0.0
flask-cors>=5.0.0
google-genai>=0.8.0
pinecone>=5.0.0
python-dotenv>=1.0.1
